Conflicts:
	CommSync/CommSync/Base.lproj/Main.storyboard

+++ CommSync/CommSync/CSAvatarSelectionView.m
@@ -38,7 +38,6 @@
     
     _app = (AppDelegate *)[[UIApplication sharedApplication] delegate];
     _sessionManager = _app.globalSessionManager;
-   // array = @[@"Avatar0", @"Avatar1", @"Avatar2", @"Avatar3", @"Avatar4", @"Avatar5", @"Avatar6", @"Avatar7", @"Avatar8", @"Avatar9", @"Avatar10", @"Avatar11", @"Avatar12", @"Avatar13", @"Avatar14", @"Avatar15", @"Avatar16", @"Avatar17", @"Avatar18", @"Avatar19", @"Avatar20", @"Avatar21", @"Avatar22", @"Avatar23", @"Avatar24", @"Avatar25", @"Avatar26", @"Avatar27" ];
 }
 
 
@@ -65,7 +64,7 @@
     {
         cell = [[CSAvatarCell alloc] init];
     }
-    NSString* avatar = [NSString stringWithFormat:@"Avatar %lu",indexPath.row + 1];
+    NSString* avatar = [NSString stringWithFormat:@"Avatar %lu",(unsigned long)indexPath.row + 1];
     [cell.avatarImage setImage:[UIImage imageNamed:avatar]];
     
     [cell.layer setCornerRadius:50.0f];
